[chore] Move parts of receiver to internal package and reexport
